디버깅코드 추가

+++ upload.php
@@ -3,7 +3,7 @@
         $id = $_POST['id'];
         $result = 0;
         $file_path = './profile_image/'.$id.'.jpg';
-        if(is_uploaded_file($_FILES['file']['tmp_name'])) {
+        if(file_exists($_FILES['file']['tmp_name']) && is_uploaded_file($_FILES['file']['tmp_name'])) {
             $result = 1;
             if(move_uploaded_file($_FILES['file']['tmp_name'], $file_path)) {
                 $result = 200;
